refactor: removing unnecessary folders, and updating the docker-compose file

+++ src/database/migrations/1661868853626-AlterTableUserDeleteUserName.ts
@@ -4,7 +4,10 @@ export class AlterTableUserDeleteUserName1661868853626
     implements MigrationInterface
 {
     public async up(queryRunner: QueryRunner): Promise<void> {
-        await queryRunner.dropColumn("users", "username");
+        const username = await queryRunner.hasColumn("users", "username");
+        if (username) {
+            await queryRunner.dropColumn("users", "username");
+        }
     }
 
     public async down(queryRunner: QueryRunner): Promise<void> {
